requirements.txt
# data

# Deep learning
torch~=2.0.1
torchdata~=0.6.1
zen3geo~=0.6.2
timm

# local package
-e .

# external requirements
# awscli
pandas~=2.0.3
geopandas~=0.13.2
src~=0.1.0
requests~=2.27.1
tqdm~=4.64.1
setuptools~=68.0.0
Shapely~=1.8.5.post1
rioxarray~=0.14.1
xarray~=2023.7.0
pytest~=7.4.0
hydra-core
xbatcher
pytest-mock